+++ src/components/common/login/Splash/index.jsx
@@ -1,0 +1,13 @@
+import logoImg from '../../../assets/full-logo.png';
+import { SplashRowBox } from './index.style';
+
+function Splash({ isActive }) {
+  return (
+      <SplashRowBox isActive={isActive}>
+          <img className='img-logo' src={logoImg} alt='' />
+          <h1 className='tit-app'>데브타운</h1>
+      </SplashRowBox>
+  );
+}
+
+export default Splash
