+++ src/components/home/HomeFeed/index.style.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const HomeFollowerUl = styled.ul`
+    margin: 20px auto;
+    width: 390px;
+`;
+
+export const HomePostItem = styled.li`
+    margin-bottom: 20px;
+`;
