+++ src/components/home/HomeNoFollower/index.style.js
@@ -1,0 +1,25 @@
+import styled from 'styled-components';
+
+export const NoFollowerMainBox = styled.div`
+    width: 100vw;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    justify-content: center;
+    align-items: center;
+    background-color: #fff;
+    padding-top: 220px;
+`;
+
+export const NoFollowerImg = styled.img`
+    width: 100px;
+    height: 100px;
+`;
+
+export const SearchUserParagraph = styled.p`
+    font-family: 'Spoqa Han Sans Neo';
+    font-size: 14px;
+    font-weight: 400;
+    line-height: 14px;
+    text-align: center;
+`;
