+++ src/pages/HomePage.jsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import TabMenu from '../components/common/TabMenu'
 import { TopMainNav } from '../components/common/nav'
-import { HomeNoFollower } from '../components/home/HomeNoFollower'
+import HomeNoFollower from '../components/home/HomeNoFollower'
 import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { homeFeedAction } from '../redux/actions/homeFeedAction'
-import HomeFeedBox from '../components/home/HomeFeedBox'
+import HomeFeedBox from '../components/home/HomeFeed'
 
 function HomePage() {
     const dispatch = useDispatch();
